Convert block numbers to uint64 for FreezeVoting contracts

+++ contracts/ERC20FreezeVoting.sol
@@ -30,12 +30,12 @@ contract ERC20FreezeVoting is BaseFreezeVoting {
         (
             address _owner,
             uint256 _freezeVotesThreshold,
-            uint256 _freezeProposalPeriod,
-            uint256 _freezePeriod,
+            uint64 _freezeProposalPeriod,
+            uint64 _freezePeriod,
             address _votesERC20
         ) = abi.decode(
                 initializeParams,
-                (address, uint256, uint256, uint256, address)
+                (address, uint256, uint64, uint64, address)
             );
 
         __Ownable_init();
@@ -56,7 +56,7 @@ contract ERC20FreezeVoting is BaseFreezeVoting {
         if (block.number > freezeProposalCreatedBlock + freezeProposalPeriod) {
             // create a new freeze proposal and set total votes to msg.sender's vote count
 
-            freezeProposalCreatedBlock = block.number;
+            freezeProposalCreatedBlock = uint64(block.number);
 
             userVotes = votesERC20.getPastVotes(
                 msg.sender,
